Move java linter function
